Deleted the trivial files
