--- app/dashboard/customers/page.tsx
@@ -1,3 +1,0 @@
-export default function Page() {
-  return <p>Customers Page</p>;
-}
