--- app/dashboard/invoices/page.tsx
@@ -1,3 +1,0 @@
-export default function Page() {
-  return <p>Invoices Page</p>;
-}
