--- app/dashboard/page.tsx
@@ -1,6 +1,0 @@
-import ChatInput from './userInputField'; 
-
-function ChatPage() {
-  return <div><ChatInput/></div>; 
-} 
-export default ChatPage;
