--- app/dashboard/userInputField.tsx
@@ -1,99 +1,0 @@
-"use client"; 
-
-import { useState, useRef } from "react";
-import {
-  ListBulletIcon,
-  PaperClipIcon,
-  ItalicIcon,
-  Bars3Icon, 
-  BoldIcon, 
-  NumberedListIcon, 
-} from '@heroicons/react/24/outline'; 
-export default function ChatInput() {
-  const [message, setMessage] = useState(""); 
-  const textAreaRef = useRef(null); // Reference to track cursor position 
-  const [setIsEmpty] = useState(true); 
-  const isEmpty = message.trim() === ""; 
-
-  // Function to apply real formatting (bold, italic, lists)
-  const applyFormatting = (command) => {
-    if (!textAreaRef.current) return;
-    document.execCommand(command, false); 
-  };
-
-// Handle text input manually
-  const handleInput = () => {
-    setMessage(textAreaRef.current.innerHTML); // Save the formatted text
-  }; 
-
-  return ( 
-    <div className="fixed bottom-0 left-[130px] w-full  p-3">
-      {/* Top Action Buttons */}
-      <div className="flex space-x-2 mb-2 max-w-2xl mx-auto">
-        <button className="bg-purple-100 text-purple-700 px-3 py-1 rounded-md text-sm">Summarize</button>
-        <button className="bg-green-100 text-green-700 px-3 py-1 rounded-md text-sm">New Idea</button>
-        <button className="bg-orange-100 text-orange-700 px-3 py-1 rounded-md text-sm">New Analytics</button>
-        <button className="bg-pink-100 text-pink-700 px-3 py-1 rounded-md text-sm">Writing Inspection</button>
-        <button className="bg-gray-100 text-gray-700 px-3 py-1 rounded-md text-sm">Uploading a document</button>
-      </div>
-
-      {/* Chat Input Box */}
-      <div className="flex flex-col max-w-7xl mx-auto border rounded-lg p-2 shadow-sm bg-gray-50">
-      
-      {/* Formatting Buttons */} 
-      <div className="flex space-x-2 mb-2">
-          <button
-            className="p-2 hover:bg-gray-200 rounded-md flex items-center"
-            onClick={() => applyFormatting("bold")}
-          >
-            <BoldIcon className="h-5 w-5 text-gray-600" />
-          </button>
-
-          <button
-            className="p-2 hover:bg-gray-200 rounded-md flex items-center"
-            onClick={() => applyFormatting("italic")}
-          >
-            <ItalicIcon className="h-5 w-5 text-gray-600" />
-          </button>
-
-          <button
-            className="p-2 hover:bg-gray-200 rounded-md flex items-center"
-            onClick={() => applyFormatting("insertOrderedList")}
-          >
-            <NumberedListIcon className="h-5 w-5 text-gray-600" />
-          </button>
-
-          <button
-            className="p-2 hover:bg-gray-200 rounded-md flex items-center"
-            onClick={() => applyFormatting("insertUnorderedList")}
-          >
-            <ListBulletIcon className="h-5 w-5 text-gray-600" />
-          </button>
-        </div>
-
-      
-
-      {/* Text Input */} 
-        <div
-          ref={textAreaRef}
-          contentEditable={true}
-          className={`border rounded-md px-3 py-3 h-24 overflow-auto focus:outline-none relative before:absolute before:left-3 before:top-3 before:text-gray-400 before:pointer-events-none bg-white ${
-            isEmpty ? "before:content-['I_am_thinking_about...']" : "before:content-[''] "
-          }`}
-          onInput={handleInput} // Track user input
-          suppressContentEditableWarning={true} // Prevents React warnings
-        ></div>
-
-        {/* Attachment & Extra Actions */} 
-        <button className="p-2 hover:bg-gray-200 rounded-md flex items-center">
-          <PaperClipIcon className="h-5 w-5 text-gray-600" />
-        </button> 
-      </div>
-
-      {/* Footer Note */}
-      <p className="text-center text-gray-500 text-xs mt-2">
-        Academico.ai can make mistakes, so double-check it.
-      </p>
-    </div> 
-  );
-}
