--- app/ui/acme-logo.tsx
@@ -1,13 +1,0 @@
-import { UserCircleIcon } from '@heroicons/react/24/outline';
-import { lusitana } from '@/app/ui/fonts';
-
-export default function AcmeLogo() {
-  return (
-    <div
-      className={`${lusitana.className} flex flex-row items-center leading-none text-white`}
-    >
-      <UserCircleIcon className="h-12 w-12 rotate-[15deg]" />
-      <p className="text-[44px]">Acme</p>
-    </div>
-  );
-}
